Made scenes easier and renamed files for consistency
I renamed the question controllers and scenes to be consistently named.

Initializing scenes can be done by adding a pair with the filename and controller to the list of pairs in QuizzzClient. In sceneCtrl you can then easily get the Scene and corresponding Controller from a hashmap by using the filename without "Scene.fxml" as a key, so "MainScreenScene.fxml" -> "MainScreen"