Refactored question controllers
I have refactored the question controllers so all of the common code (e.g. initializing the scene, starting the timer) are in one class.This way we don't need to have a bunch of duplicate code in each controller. The controllers for each specific question then extend this class so they keep the same functionality. I also allowed protected fields in the checkstyle.