Unified multiplayer/singleplayer leaderboard methods
Instead of having a singleplayer and multiplayer table in the DB, there is one leaderboard table that stores the roomId. When retrieving the roomId needs to be specified and a query will be made to only get the entries with that roomId. The multiplayer entries are deleted if another game is started with the same roomId, this should eventually be changed to be when the game ends.